Create auto remediation script
I observed very worrisome behaviour when some of the healthchecks got fallen. Then
the new arrived ec2 instances weren't able to run the app: into the instances are missing the start scripts and all needed requisites which are usually coming as a codedeploy artifacts.
Thats why I created script to lookup for these artifacts (last stable build) and to bring them on the instance and then to run the service.